Start of the piano IoT project (without the AI)
need to add the code so that the buzzer on alivecode is a replication of the buzzer in real life and that there's log messages on alivecode. The log messages should display which note has been clicked (ex. if you clicked 'do' on the keyboard, it should show something like 'do has been clicked')